Rename src/ to app/
